# docs/specs/openapi/control-plane.yaml
openapi: 3.1.0
info:
  title: Maestro Control Plane API
  version: 3.0.0
  license:
    name: Internal Proprietary
    identifier: LicenseRef-Internal-Proprietary
  description: >-
    Maestro v3 central control-plane contract. Project scope, role and delegated
    session are derived from the authenticated server context and resource path;
    request bodies never establish authorization scope. The API does not expose
    source-branch push, protected-branch push or merge operations.
servers:
  - url: /api/v3
security:
  - oidcBearer: []
...
paths:
...
  /gitlab/instances:
    get:
      tags: [GitLab]
      operationId: listGitLabInstances
      summary: List configured GitLab instances without exposing Secrets
      description: Lists sanitized GitLab instance configuration visible to the platform administrator.
      responses:
        '200':
          description: Configured GitLab instances without secrets
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/GitLabInstance'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [GitLab]
      operationId: createGitLabInstance
      summary: Configure an approved GitLab instance
      description: Creates an audited HTTPS GitLab instance configuration without returning Secrets.
      x-maestro-permission: gitlab_instance.configure
      x-maestro-audit-event: gitlab.instance_created
      x-maestro-idempotency: required-header
      x-maestro-concurrency: if-none-match-required
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfNoneMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GitLabInstanceCreate'}
      responses:
        '201':
          description: GitLab instance created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GitLabInstance'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
...
  /projects/{project_id}/gitlab/merge-requests/{merge_request_iid}:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/MergeRequestIid'}
    get:
      tags: [GitLab]
      operationId: getMergeRequestProjection
      summary: Read the reconciled GitLab merge request projection
      description: Returns the cached projection last verified against GitLab facts.
      responses:
        '200':
          description: Read-only reconciled MR and Pipeline projection
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MergeRequestProjection'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /projects/{project_id}/gitlab/merge-requests/{merge_request_iid}/reconcile:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/MergeRequestIid'}
    post:
      tags: [GitLab]
      operationId: reconcileMergeRequest
      summary: Request read-only reconciliation with GitLab facts
      description: Refreshes read models only; it cannot merge or push any branch.
      x-maestro-permission: gitlab.reconcile
      x-maestro-audit-event: gitlab.merge_request_reconcile_requested
      x-maestro-idempotency: required-header
      x-maestro-concurrency: if-match-required
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReconcileRequest'}
      responses:
        '202': {$ref: '#/components/responses/AcceptedOperation'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
  /webhooks/dead-letters/{inbox_id}/replay:
    parameters:
      - {$ref: '#/components/parameters/InboxId'}
    post:
      tags: [GitLab]
      operationId: replayDeadLetter
      summary: Dual-approved replay of one quarantined webhook delivery
      description: >-
        Re-queues a dead-letter Inbox row under its ORIGINAL event identity:
        no re-verification bypass and no side-effect duplication. The approver
        identity is the authenticated principal; it must differ from
        requested_by (runbook section 9 dual-person control). The requeue, the
        approval and the audit row land in one transaction.
      x-maestro-permission: gitlab.reconcile
      x-maestro-audit-event: webhook.dead_letter.replayed
      x-maestro-idempotency: quarantined-row-state-collapse
      x-maestro-concurrency: single-row-transactional-guard
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DeadLetterReplay'}
      responses:
        '200':
          description: Quarantined row re-queued under its original identity
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DeadLetterReplayResult'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
  /projects/{project_id}/audit-export:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    get:
      tags: [Audit]
      operationId: exportAuditChain
      summary: Export one project's append-only audit chain slice
      description: >
        Walks the immutable audit_events table in seq order and returns the
        frozen audit-export wire: ordered entries with deterministic uuid event
        ids, per-entry and rolling chain digests, and the redaction record.
        Fields outside the redaction allowlist (token hashes, free-text
        reasons) never appear.
      x-maestro-permission: audit.export
      x-maestro-audit-event: audit.exported
      parameters:
        - name: from_seq
          in: query
          required: true
          description: First audit sequence number in the slice (inclusive).
          schema: {type: integer, format: int64, minimum: 1}
        - name: to_seq
          in: query
          required: true
          description: Last audit sequence number in the slice (inclusive).
          schema: {type: integer, format: int64, minimum: 1}
      responses:
        '200':
          description: The audit chain export with per-entry digests
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuditChainExport'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /projects/{project_id}/audit-export/verify:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    post:
      tags: [Audit]
      operationId: verifyAuditChain
      summary: Recompute the audit chain and compare claimed digests
      description: >
        The auditor's tamper check over an export: recomputes the chain for
        the range and compares it against the claimed per-entry digests. The
        If-Match precondition carries the chain digest of the export being
        verified.
      x-maestro-permission: audit.export
      x-maestro-audit-event: audit.exported
      x-maestro-idempotency: required-header
      x-maestro-concurrency: if-match-required
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuditVerifyRequest'}
      responses:
        '200':
          description: The claimed digests match the recomputed chain
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuditVerifyResult'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
  /projects/{project_id}/slo-snapshot:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    get:
      tags: [Audit]
      operationId: getSLOSnapshot
      summary: Evaluate one project's SLO snapshot
      description: >
        Evaluates the frozen slo-status wire at request time: declared
        objectives from the project's telemetry aggregates, the deployment
        objectives (RPO, RTO, backup success rate) from the backup ledger,
        availability from the configured SLI metrics. Missing availability
        telemetry fails closed with 503 — the snapshot never invents a
        number. Absent configuration keeps the endpoint unexposed.
      x-maestro-permission: project.read
      responses:
        '200':
          description: The SLO snapshot (the canonical slo-status schema)
          content:
            application/json:
              schema: {$ref: '../schemas/slo-status.schema.json'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /projects/{project_id}/quality-policy:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    get:
      tags: [Quality]
      operationId: getEffectiveQualityPolicy
      summary: Read the effective inherited Quality Policy
      description: Returns the effective company and Project policy after monotonic inheritance.
      responses:
        '200':
          description: Effective inherited policy and provenance
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EffectiveQualityPolicy'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Quality]
      operationId: putProjectQualityPolicy
      summary: Strengthen the Project Quality Policy
      description: Project policy may only strengthen its company baseline.
      x-maestro-permission: project_policy.strengthen
      x-maestro-audit-event: quality.policy_changed
      x-maestro-idempotency: required-header
      x-maestro-concurrency: if-match-or-if-none-match-required
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatchOptional'}
        - {$ref: '#/components/parameters/IfNoneMatchOptional'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/QualityPolicyPut'}
      responses:
        '200':
          description: Policy replaced and versioned
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EffectiveQualityPolicy'}
        '201':
          description: Project policy created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EffectiveQualityPolicy'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
  /projects/{project_id}/work-items/{work_item_id}/gates:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/WorkItemId'}
    get:
      tags: [Quality]
      operationId: listWorkItemGates
      summary: List exact-SHA Gate snapshots for a WorkItem
      description: Lists Gate states bound to the WorkItem's exact source, target, and policy tuple.
      responses:
        '200':
          description: Gate snapshots for the exact source/target/policy tuple
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Gate'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /projects/{project_id}/work-items/{work_item_id}/evidence:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/WorkItemId'}
    get:
      tags: [Quality]
      operationId: listWorkItemEvidence
      summary: List immutable Evidence for a WorkItem
      description: Lists immutable Evidence records authorized for the WorkItem scope.
      responses:
        '200':
          description: Immutable Evidence, including authority and exact SHA tuple
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Evidence'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /projects/{project_id}/work-items/{work_item_id}/waivers:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/WorkItemId'}
    get:
      tags: [Quality]
      operationId: listWorkItemWaivers
      summary: List the Waiver lifecycle queue for a WorkItem
      description: >-
        Read-only queue view for HITL review: every requested, approved or
        terminal Waiver with its live remaining validity window. Waiver
        transitions stay on the frozen write endpoints.
      responses:
        '200':
          description: Waiver queue entries with remaining validity windows
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - {$ref: '#/components/schemas/Waiver'}
                    - type: object
                      additionalProperties: false
                      required: [remaining_seconds]
                      properties:
                        remaining_seconds: {type: integer, format: int64, minimum: 0}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /projects/{project_id}/gates/{gate_id}/waivers:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/GateId'}
    post:
      tags: [Quality]
      operationId: requestGateWaiver
      summary: Request a time-limited waiver for one eligible Gate tuple
      description: Requests a maximum seven-day waiver bound to one MR, SHA and check.
      x-maestro-permission: waiver.request
      x-maestro-audit-event: quality.waiver_requested
      x-maestro-idempotency: required-header
      x-maestro-concurrency: if-match-required
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WaiverRequest'}
      responses:
        '201':
          description: Waiver request awaiting a distinct approver
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Waiver'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
  /projects/{project_id}/waivers/{waiver_id}/approve:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/WaiverId'}
    post:
      tags: [Quality]
      operationId: approveGateWaiver
      summary: Independently approve an eligible Gate Waiver
      description: Approves a Waiver only when separation-of-duties and validity rules pass.
      x-maestro-permission: waiver.approve
      x-maestro-audit-event: quality.waiver_approved
      x-maestro-idempotency: required-header
      x-maestro-concurrency: if-match-required-and-separation-of-duties
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReasonRequest'}
      responses:
        '200':
          description: Approved waiver
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Waiver'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
  /projects/{project_id}/waivers/{waiver_id}/revoke:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/WaiverId'}
    post:
      tags: [Quality]
      operationId: revokeGateWaiver
      summary: Revoke an active Gate Waiver
      description: Revokes an active Waiver with optimistic concurrency and audit recording.
      x-maestro-permission: waiver.revoke
      x-maestro-audit-event: quality.waiver_revoked
      x-maestro-idempotency: required-header
      x-maestro-concurrency: if-match-required
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReasonRequest'}
      responses:
        '200':
          description: Revoked waiver
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Waiver'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookToken:
      type: apiKey
      in: header
      name: X-Gitlab-Token
  parameters:
...
    ProjectId:
      name: project_id
      in: path
      required: true
      description: Resource locator only; authorization scope is server-derived.
      schema: {$ref: '#/components/schemas/Identifier'}
...
    InboxId:
      name: inbox_id
      in: path
      required: true
      schema: {type: string, format: uuid}
    MergeRequestIid:
      name: merge_request_iid
      in: path
      required: true
      schema: {type: integer, minimum: 1}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string, pattern: '^"[1-9][0-9]*"$'}
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: true
      schema: {type: string, const: '*'}
    IfMatchOptional:
      name: If-Match
      in: header
      required: false
      schema: {type: string, pattern: '^"[1-9][0-9]*"$'}
...
  headers:
    ETag:
      description: Strong aggregate version ETag.
      schema: {type: string, pattern: '^"[1-9][0-9]*"$'}
  schemas:
...
    GitLabInstance:
      type: object
      additionalProperties: false
      required: [id, base_url, tls_verify, state, version]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        base_url: {type: string, format: uri, pattern: '^https://'}
        tls_verify: {type: boolean, const: true}
        state: {type: string, enum: [active, degraded, revoked]}
        version: {$ref: '#/components/schemas/ResourceVersion'}
    GitLabInstanceCreate:
      type: object
      additionalProperties: false
      required: [base_url, bot_secret_ref, webhook_secret_ref]
      properties:
        base_url: {type: string, format: uri, pattern: '^https://'}
        bot_secret_ref: {type: string, minLength: 1, maxLength: 255}
        webhook_secret_ref: {type: string, minLength: 1, maxLength: 255}
...
    MergeRequestProjection:
      type: object
      additionalProperties: false
      required: [iid, state, source_sha, target_sha, observed_at]
      properties:
        iid: {type: integer, minimum: 1}
        state: {type: string, enum: [opened, closed, merged]}
        source_sha: {$ref: '#/components/schemas/GitCommitSha'}
        target_sha: {$ref: '#/components/schemas/GitCommitSha'}
        pipeline_id: {type: integer, minimum: 1}
        observed_at: {$ref: '#/components/schemas/Timestamp'}
        stale: {type: boolean}
...
    ReconcileRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: {type: string, minLength: 8, maxLength: 2000}
        force_provider_read: {type: boolean, default: false}
...
    Evidence:
      $ref: '../schemas/evidence.schema.json'
...
    EffectiveQualityPolicy:
      type: object
      additionalProperties: false
      required: [project_id, effective_policy, policy_digest, provenance]
      properties:
        project_id: {$ref: '#/components/schemas/Identifier'}
        effective_policy: {$ref: '#/components/schemas/QualityPolicy'}
        policy_digest:
          type: string
          pattern: '^sha256:[a-f0-9]{64}$'
        provenance:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [scope, policy_id, version]
            properties:
              scope: {type: string, enum: [company, project, task]}
              policy_id: {type: string, minLength: 1, maxLength: 64}
              version: {type: string, pattern: '^[0-9]+\.[0-9]+\.[0-9]+$'}
              digest: {type: string, pattern: '^sha256:[a-f0-9]{64}$'}
...
    DeadLetterReplay:
      type: object
      additionalProperties: false
      required: [requested_by, reason]
      properties:
        requested_by:
          type: string
          minLength: 1
          maxLength: 200
          description: The principal who diagnosed and requested the replay.
        approved_by:
          type: string
          minLength: 1
          maxLength: 200
          description: >-
            Optional; when present it must name the authenticated principal —
            the server always derives the approver from the credential, never
            the body.
        reason:
          type: string
          minLength: 16
          maxLength: 2000
          description: Substantive justification recorded with the audit row.
    DeadLetterReplayResult:
      type: object
      additionalProperties: false
      required: [inbox_id, replayed]
      properties:
        inbox_id: {type: string, format: uuid}
        replayed: {type: boolean, const: true}
    AuditChainExport:
      type: object
      additionalProperties: false
      required: [schema_version, export_id, project_id, range, entries, chain_digest]
      properties:
        schema_version: {type: string, const: "3.0"}
        export_id: {type: string, format: uuid}
        project_id: {type: string, format: uuid}
        range:
          type: object
          additionalProperties: false
          required: [from_seq, to_seq]
          properties:
            from_seq: {type: integer, format: int64, minimum: 1}
            to_seq: {type: integer, format: int64, minimum: 1}
        entries:
          type: array
          items: {$ref: '#/components/schemas/AuditExportEntry'}
        chain_digest:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        redaction:
          type: object
          additionalProperties: false
          properties:
            policy_version: {type: string, minLength: 1, maxLength: 64}
            fields_masked:
              type: array
              uniqueItems: true
              items: {type: string, minLength: 1, maxLength: 64}
...
    AuditVerifyRequest:
      type: object
      additionalProperties: false
      required: [from_seq, to_seq, claimed_digests]
      properties:
        from_seq: {type: integer, format: int64, minimum: 1}
        to_seq: {type: integer, format: int64, minimum: 1}
        claimed_digests:
          type: array
          items: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}
    AuditVerifyResult:
      type: object
      additionalProperties: false
      required: [verified]
      properties:
        verified: {type: boolean}
...
  responses:
    AcceptedOperation:
      description: Durable asynchronous operation accepted
      content:
        application/json:
          schema: {$ref: '#/components/schemas/OperationAccepted'}
    BadRequest:
      description: Malformed request (`INVALID_REQUEST`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthorized:
      description: Authentication failed (`AUTHENTICATION_REQUIRED`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: Authenticated principal lacks permission (`PERMISSION_DENIED`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: Resource absent or hidden by project visibility (`RESOURCE_NOT_FOUND`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: State, idempotency hash or uniqueness conflict (`STATE_CONFLICT`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    PreconditionFailed:
      description: ETag no longer matches (`VERSION_CONFLICT`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
...
    Unprocessable:
      description: Semantic validation or immutable rule failed (`VALIDATION_FAILED`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    PreconditionRequired:
      description: Required conditional header missing (`PRECONDITION_REQUIRED`)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}
    ServiceUnavailable:
      description: Dependency unavailable or fail-closed safety mode
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Error'}